refactor(watchlists): extract Console handoff from the screen shell
